refactor(ui): move appearance resolution out of model.go
Theme, accent, background ownership, and colour mode resolve independently and
are previewed and persisted through one modal. Keeping that in model.go left
the appearance rules interleaved with service lifecycle code across 1000 lines.

Declarations move unchanged.
